fix(logger): Change timestamp type from uint16_t to uint32_t in loggerLog function

+++ logger.h
@@ -41,7 +41,7 @@
 
     LoggerStatus loggerLog(
         Logger *logger, 
-        uint16_t timestamp, 
+        uint32_t timestamp, 
         LogLevel level, 
         const char *message
     );
